ERA-13478: move the account-linking gate client from ducks to utils
checkAccountLinked touches no Redux (no store read, no dispatch — it just returns the classified gate response), so it doesn't belong in a duck. With the finalizing-flag reducer gone, the module has no store state at all; move it (and its test) to src/utils/account-linking.js and repoint Auth0TokenManager's imports. Addresses PR #1603 review.
